Minor tweaks from self-review
